게시물 삭제 기능 추가

+++ src/main/webapp/jsp/usr/article/list.jsp
@@ -20,10 +20,10 @@
             <% for ( ArticleDto article : articles ) { %>
             <li class="flex">
                 <a class="w-[60px] hover:underline hover:text-[blue]" href="/usr/article/detail/free/<%=article.getId()%>"><%=article.getId()%></a>
-                <a class="w-[60px] hover:underline hover:text-[blue]" href="/usr/article/detail/free/<%=article.getId()%>"><%=article.getTitle()%></a>
                 <!-- flex-grow : 성장성 1 -->
-                <a class="flex-grow hover:underline hover:text-[blue]" href="/usr/article/modify/free/<%=article.getId()%>">수정</a>
-                <a class="w-[100px] hover:underline hover:text-[blue]" href="/usr/article/delete/free/<%=article.getId()%>">삭제</a>
+                <a class="flex-grow hover:underline hover:text-[blue]" href="/usr/article/detail/free/<%=article.getId()%>"><%=article.getTitle()%></a>
+<%--                <a class="flex-grow hover:underline hover:text-[blue]" href="/usr/article/modify/free/<%=article.getId()%>">수정</a>--%>
+                <a onclick="if ( !confirm('정말로 삭제하시겠습니까?') ) return false;" class="w-[100px] hover:underline hover:text-[blue]" href="/usr/article/delete/free/<%=article.getId()%>">삭제</a>
             </li>
             <% } %>
         </ul>
